+++ src/shared/models/post/post.model.ts
@@ -1,0 +1,7 @@
+import { ID } from 'scalars';
+
+export class PostModel {
+  public id: ID;
+  public title?: string;
+  public content: string;
+}