+++ src/user/user.controller.spec.ts
@@ -1,8 +1,8 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { UserController } from './user.controller';
 import { UserService } from './user.service';
+import { UserController } from './user.controller';
 import { UserRepository } from './user.repository';
-import { CreateUserDto } from 'src/shared/models/create-user.dto';
+import { Test, TestingModule } from '@nestjs/testing';
+import { CreateUserDto } from '../shared/models/user/create-user.dto';
 
 describe('UserController', () => {
   let userControler: UserController;