+++ src/user/user.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { ID } from 'scalars';
-import { User } from '../shared/models/user.model';
+import { User } from '../shared/models/user/user.model';
 import { UserRepository } from './user.repository';
-import { CreateUserDto } from '../shared/models/create-user.dto';
-import { UpdateUserDto } from '../shared/models/update-user.dto';
+import { CreateUserDto } from '../shared/models/user/create-user.dto';
+import { UpdateUserDto } from '../shared/models/user/update-user.dto';
 
 @Injectable()
 export class UserService {